user_app/announcements/serializers.py
from rest_framework import serializers
from shared.models import Announcement


class AnnouncementSerializer(serializers.ModelSerializer):
    class Meta:
        model = Announcement
        fields = "__all__"
