requirements.txt
asgiref==3.8.1
black==24.10.0
boto3==1.21.10
botocore==1.24.46
click==8.1.7
Django==4.2.16
django-cors-headers==4.6.0
django-hosts==6.0
django-ses==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
isort==5.13.2
jmespath==0.10.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.20
